Promote reuse of test runner support code
